Change SQLTemplate update method (for suffix)
SQLTemplate file updates (i.e. iterations control related queries)
that are to be appended to the SQL file are now applied before the
latest 'select' string occurence. This happens due to the rule that
all template SQL files need to have an output, thus by appending to
the file, the output was 'blocked'.