Add local (and federated) tests to CB
Currently only the local tests pass and only if I comment lines 64-65
in calibration_belt.py. The reason for that is that the tests Stefano
sent me have very few data each and they correspond to the regime the
algorithm shouldn't run normally.

Secod, the federated tests fail for a much more important reason. The
scarcity of data in these tests revealed a bug plaguing all algorithms.
If all levels of a categorical variable happen not to be represented
in all local dbs, they way we exctract data (sqlalchemy+patsy) results
in local tables where the schemas do not agree.

There should be a general solution for this in the future.