Refactoring of descriptive statistics (+bug fix)
Descriptive stats algorithm now outputs two views of the variables
statistcs.
One on a per variable basis, where statistics are presented
for each variable individually, and one for the so-called model,
i.e. the selection of variables on which some algorithm will run.
The second differs from the first on the fact that only the
intersection of non-empty rows is kept, all other rows being discarded.
This can, potentially, result in very different number of datapoints
for the two views.