Change the way LogReg decides binary target values
The way it was done resulted in a subtle bug where sometimes the target
levels where not assigned consistently on different nodes.

This is now fixed: The user inputs the levels he wishes to assign to
positive and negative outcomes. This also eliminates the need to use
filters to handle multilevel variables. Only two levels are always
filter from within the algorithm now (method: keep_levels).