Implementation of one-way Anova
Doesn't work correctly yet! Midway refactoring: removing the possibility
to run anova on multiple covariables, as this would be best implemented
using a) the formula and b) a linear regression model as a first step.
Since we don't yet have formulae and algorithm pipelines there is no
point in implementing the multivariate anova version.