Add new formula functionality
Algorithms now accept a new formula argument. This arg expects a
description of a formula in Wilkinson notation in JSON format. From the
description an actual string formula is constructed and passed to the
patsy library from which the algorithm's design matrices are
constructed. The rest of the execution remains unchanged.

Currently only a subset of the full Wilkinson notation is implemented.
Interactions are up to 3 terms and singleton terms have a limited number of
transformations allowed. These are nop, log, exp, center, standardize,
mul and div for numerical variables.