Move coding processing in parameters.py
Currently logistic_regression tests with formula are failing. There is a
subtle interaction between the formula coding of categorical
variables and the the add_missing_levels method. The reason is that
add_missing_levels is based on the known enumerations of categorical
vars found the in the CDEs but the codings supported by the formula
might add different enums to categorical vars.